move generic models out of FILER; also create GenomicsDB directory placeholders
