+++ api/response_models/__init__.py
@@ -1,3 +1,7 @@
 from .base_models import PagedResponseModel, BaseResponseModel, SerializableModel, RequestDataModel, GenericDataModel
 from .genome_browser import GenomeBrowserConfig, GenomeBrowserConfigResponse, GenomeBrowserExtendedConfigResponse, GenomeBrowserExtendedConfig
 from .data_models import BEDFeature, BEDResponse
+from .biosample_characteristics import BiosampleCharacteristics
+from .experimental_design import ExperimentalDesign
+from .provenance import Provenance
+
