+++ api/response_models/genome_browser.py
@@ -9,10 +9,11 @@ from niagads.utils.list import find
 from api.common.constants import JSON_TYPE
 from api.common.enums import OnRowSelect, ResponseFormat
 from api.common.formatters import id2title
-from api.routers.filer.models.biosample_characteristics import BiosampleCharacteristics
-from api.routers.filer.models.experimental_design import ExperimentalDesign # TODO: possibly move to API base_models
+
 
 from .base_models import BaseResponseModel, RowModel
+from .biosample_characteristics import BiosampleCharacteristics
+from .experimental_design import ExperimentalDesign 
 
 class GenomeBrowserConfig(RowModel,SQLModel):
     track_id: str = Field(alias='id')
