+++ api/response_models/provenance.py
@@ -1,11 +1,11 @@
 
 from datetime import datetime
 from typing import Optional
-from pydantic import BaseModel
-from api.response_models.base_models import SerializableModel
+
+from .base_models import SerializableModel
 
 
-class Provenance(SerializableModel, BaseModel):
+class Provenance(SerializableModel):
     data_source: Optional[str]
     data_source_version: Optional[str]
     download_url: Optional[str]
